fix preload running twice when logout/login
