+++ das_client/app/integration_test/mocks/mock_authenticated_scope.dart
@@ -5,10 +5,8 @@ import 'package:app/di/scopes/authenticated_scope.dart';
 import 'package:app/provider/ru_feature_provider.dart';
 import 'package:formation/component.dart';
 import 'package:logging/logging.dart';
-import 'package:preload/component.dart';
 
 import 'mock_formation_repository.dart';
-import 'mock_preload_repository.dart';
 import 'mock_ru_feature_provider.dart';
 
 final _log = Logger('MockAuthenticatedScope');
@@ -31,15 +29,12 @@ class MockAuthenticatedScope extends AuthenticatedScope {
     getIt.registerMqttAuthProvider();
     getIt.registerMqttService();
     getIt.registerHttpClient();
-    getIt.registerSferaLocalRepo();
     getIt.registerSferaRemoteRepo();
     getIt.registerSettingsRepository();
     if (e2e) {
-      getIt.registerPreloadRepository();
       getIt.registerRuFeatureProvider();
       getIt.registerFormationRepository();
     } else {
-      _registerMockPreloadRepository();
       _registerMockRuFeaturesProvider();
       _registerMockFormationRepository();
     }
@@ -53,9 +48,5 @@ class MockAuthenticatedScope extends AuthenticatedScope {
 
   void _registerMockFormationRepository() {
     getIt.registerSingleton<FormationRepository>(MockFormationRepository());
-  }
-
-  void _registerMockPreloadRepository() {
-    getIt.registerSingleton<PreloadRepository>(MockPreloadRepository());
   }
 }
