+++ das_client/app/lib/di/scopes/das_base_scope.dart
@@ -1,6 +1,7 @@
 import 'package:app/brightness/brightness_manager.dart';
 import 'package:app/brightness/brightness_manager_impl.dart';
 import 'package:app/di/di.dart';
+import 'package:app/flavor.dart';
 import 'package:app/launcher/launcher.dart';
 import 'package:app/launcher/launcher_impl.dart';
 import 'package:app/pages/login/login_view_model.dart';
@@ -13,7 +14,9 @@ import 'package:battery_plus/battery_plus.dart';
 import 'package:connectivity_x/component.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logging/logging.dart';
+import 'package:preload/component.dart';
 import 'package:screen_brightness/screen_brightness.dart';
+import 'package:sfera/component.dart';
 import 'package:warnapp/component.dart';
 
 final _log = Logger('DASBaseScope');
@@ -39,6 +42,9 @@ class DASBaseScope extends DIScope {
     getIt.registerLoginViewModel();
     getIt.registerAppLinksManager();
     getIt.registerLauncher();
+    getIt.registerSferaLocalRepo();
+    getIt.registerPreloadRepository();
+
     await getIt.allReady();
   }
 }
@@ -117,5 +123,23 @@ extension BaseScopeExtension on GetIt {
   void registerLauncher() {
     _log.fine('Register Launcher');
     registerSingleton<Launcher>(LauncherImpl(userSettings: DI.get()));
+  }
+
+  void registerSferaLocalRepo() {
+    factoryFunc() {
+      _log.fine('Register sfera local repo');
+      return SferaComponent.createSferaLocalRepo();
+    }
+
+    registerLazySingleton<SferaLocalRepo>(factoryFunc);
+  }
+
+  void registerPreloadRepository() {
+    registerSingleton<PreloadRepository>(
+      PreloadComponent.createPreloadRepository(
+        sferaLocalRepo: DI.get(),
+        disablePreload: DI.get<Flavor>().disablePreload,
+      ),
+    );
   }
 }
